Create HTML skeleton with basic styling
The web app skeleton is created, namely css and js folders. Basic
styling and content has been added to the index.html to demonstrate a
working site and to view possible stylings.

The JS-side of the app is started but incomplete. It requires further
supporting markup, content, and programming to tie these together.